feat: use vector commitment methods to verify in lightclient (#303)
* chore: move tendermint spec specification from test to main


* feat: use methods to verifyMembership and verifyNonMembership from vector commitment


* fix: fix use of key in verification


---------
