+++ contracts/javascore/ibc/src/main/java/ibc/ics23/commitment/Proof.java
@@ -393,4 +393,26 @@ public class Proof {
         }
         return true;
     }
+
+    public static ProofSpec getTendermintSpec() {
+        LeafOp leafSpec = new LeafOp();
+        leafSpec.setPrefix(new byte[]{0});
+        leafSpec.setPrehashKey(HashOp.NO_HASH);
+        leafSpec.setHash(HashOp.SHA256);
+        leafSpec.setPrehashValue(HashOp.SHA256);
+        leafSpec.setLength(LengthOp.VAR_PROTO);
+
+        InnerSpec innerSpec = new InnerSpec();
+        innerSpec.setChildOrder(List.of(BigInteger.ZERO, BigInteger.ONE));
+        innerSpec.setMinPrefixLength(BigInteger.ONE);
+        innerSpec.setMaxPrefixLength(BigInteger.ONE);
+        innerSpec.setChildSize(BigInteger.valueOf(32));
+        innerSpec.setHash(HashOp.SHA256);
+
+        var tendermintSpec = new ProofSpec();
+        tendermintSpec.setLeafSpec(leafSpec);
+        tendermintSpec.setInnerSpec(innerSpec);
+        return tendermintSpec;
+    }
+
 }
